+++ src/App.vue
@@ -1,66 +1,24 @@
 <template>
   <v-app v-bind="this.$store.state.theme">
-    <v-toolbar>
-      <v-toolbar-title class="mr-5">
-        <router-link to="/">Elasticvue</router-link>
-      </v-toolbar-title>
-
-      <v-chip :color="connectionStateClass">{{connectionStateName}}</v-chip>
-
-      <v-spacer></v-spacer>
-      <connect isToolbar v-if="isConnected()" v-on:hostChanged="rerender"></connect>
-      <v-spacer></v-spacer>
-
-      <v-toolbar-items>
-        <v-btn flat to="/" exact>Home</v-btn>
-        <v-btn flat to="/indices">Indices</v-btn>
-        <v-btn flat to="/browse">Browse</v-btn>
-        <v-btn flat to="/query">Query</v-btn>
-        <v-btn flat to="/utilities">Utilities</v-btn>
-        <v-btn flat href="https://github.com/cars10/elasticvue" target="_blank">
-          <img src="../static/GitHub-Mark-Light.png" alt="GithubIcon" v-if="this.$store.state.theme.dark">
-          <img src="../static/GitHub-Mark.png" alt="GithubIcon" v-else>
-        </v-btn>
-      </v-toolbar-items>
-    </v-toolbar>
+    <app-header v-on:hostChanged="rerender"></app-header>
 
     <v-content>
       <v-container grid-list-md>
-        <content-or-connect>
+        <content-or-setup>
           <router-view v-if="renderRouterView"></router-view>
-        </content-or-connect>
+        </content-or-setup>
       </v-container>
       <snackbar></snackbar>
     </v-content>
 
-    <v-footer height="auto">
-      <v-layout row wrap py-3 px-3>
-        <v-flex xs4 align-center d-inline-flex>
-          <v-switch label="Dark theme" v-model="theme" hide-details></v-switch>
-        </v-flex>
-
-        <v-flex xs4 text-xs-center>
-          &copy;{{ new Date().getFullYear()}}<br/>
-          <v-btn @click="reset" title="Reset saved settings">Reset</v-btn>
-        </v-flex>
-
-        <v-flex xs4 text-xs-right align-center d-inline-flex>
-          <v-flex>
-            <a href="https://github.com/cars10/elasticvue" target="_blank" rel="nofollow">
-              <img src="../static/GitHub_Logo_White.png" alt="GithubLogo" v-if="this.$store.state.theme.dark">
-              <img src="../static/GitHub_Logo.png" alt="GithubLogo" v-else>
-            </a>
-          </v-flex>
-        </v-flex>
-      </v-layout>
-    </v-footer>
+    <app-footer></app-footer>
   </v-app>
 </template>
 
 <script>
-  import Connect from '@/components/shared/Connect'
-  import ContentOrConnect from '@/components/shared/ContentOrConnect'
-  import { CONNECTION_STATE_CLASSES, CONNECTION_STATE_NAMES } from './consts'
+  import AppHeader from '@/components/App/AppHeader'
+  import AppFooter from '@/components/App/AppFooter'
+  import ContentOrSetup from '@/components/ContentOrSetup'
   import Snackbar from '@/components/Snackbar'
 
   export default {
@@ -70,37 +28,18 @@
         renderRouterView: true
       }
     },
-    computed: {
-      connectionStateName () {
-        return CONNECTION_STATE_NAMES[this.$store.state.connection.status]
-      },
-      connectionStateClass () {
-        return CONNECTION_STATE_CLASSES[this.$store.state.connection.status]
-      },
-      theme: {
-        get () {
-          return this.$store.state.theme.dark
-        },
-        set (value) {
-          this.$store.commit('setTheme', value)
-        }
-      }
-    },
     methods: {
-      reset () {
-        localStorage.clear()
-        window.location.replace('/')
-      },
       rerender () {
         this.renderRouterView = false
-        this.$store.commit('resetBrowse')
+        this.$store.commit('resetSearch')
         this.$store.commit('resetConnection')
         this.$nextTick(() => (this.renderRouterView = true))
       }
     },
     components: {
-      Connect,
-      ContentOrConnect,
+      AppHeader,
+      AppFooter,
+      ContentOrSetup,
       Snackbar
     }
   }