+++ src/components/Indices/IndicesTable.vue
@@ -4,7 +4,7 @@
       <v-flex right d-inline-flex>
         <v-text-field append-icon="search"
                       v-on:keyup.esc="indicesFilter = ''"
-                      label="Filter..."
+                      label="Filter via 'column:query'"
                       name="filter"
                       id="filter"
                       v-model="indicesFilter"></v-text-field>
@@ -14,6 +14,7 @@
                   :headers="headers"
                   :items="indices"
                   :custom-sort="sortIndices"
+                  :custom-filter="callFuzzyTableFilter"
                   :search="indicesFilter"
                   :loading="loading"
                   class="table--condensed">
@@ -30,7 +31,7 @@
           <td class="text-xs-right">{{props.item['pri.store.size']}}</td>
           <td>
             <btn-group small>
-              <v-btn flat @click.native.stop="showDocuments(props.item.index)" title="Browse documents">
+              <v-btn flat @click.native.stop="showDocuments(props.item.index)" title="Search documents">
                 <v-icon>view_list</v-icon>
               </v-btn>
               <v-btn flat @click.native.stop="openIndex(props.item.index)" title="Show">
@@ -49,13 +50,14 @@
 
 <script>
   import BtnGroup from '@/components/shared/BtnGroup'
+  import { fuzzyTableFilter } from '../../helpers/filters'
 
   export default {
     name: 'IndicesTable',
     data () {
       return {
         headers: [
-          {text: 'Name', value: 'index'},
+          {text: 'Index', value: 'index'},
           {text: 'Health', value: 'health'},
           {text: 'Status', value: 'status'},
           {text: 'UUID', value: 'uuid'},
@@ -109,8 +111,8 @@
         })
       },
       showDocuments (index) {
-        this.$store.commit('setBrowseIndices', [index]) // to pre-select right index on "Browse" page
-        this.$router.push({name: 'Browse', params: {executeSearch: true}})
+        this.$store.commit('setSearchIndices', [index]) // to pre-select right index on "Search" page
+        this.$router.push({name: 'Search', params: {executeSearch: true}})
       },
       openIndex (index) {
         this.$router.push({name: 'Index', params: {index: index}})
@@ -125,6 +127,9 @@
             })
             .catch(error => this.$store.commit('setErrorState', error))
         }
+      },
+      callFuzzyTableFilter (items, search, filter, headers) {
+        return fuzzyTableFilter(items, search, filter, headers)
       }
     },
     components: {