+++ src/components/Search.vue
@@ -1,7 +1,7 @@
 <template>
   <v-card>
     <v-card-title>
-      <h1 class="headline">Browse</h1>
+      <h1 class="headline">Search</h1>
       <reload-button title="Reload indices" :action="() => $refs.indicesLoader.loadData()"></reload-button>
     </v-card-title>
     <v-divider></v-divider>
@@ -10,22 +10,28 @@
       <v-form v-on:submit.prevent="loadData">
         <v-layout row wrap>
           <v-flex lg2>
-            <v-text-field label="Query" name="Query" id="query" v-model="browseQ"></v-text-field>
+            <v-text-field label="Query" name="Query" id="query" v-model="searchQ"></v-text-field>
           </v-flex>
 
           <v-flex>
             <data-loader method="catIndices" ref="indicesLoader" renderContentWhileLoading>
               <template slot-scope="data">
-                <v-select multiple
-                          auto
-                          autocomplete
-                          label="Indices"
-                          name="Indices"
-                          id="indices"
-                          v-model="browseIndices"
-                          :items="data.body | sortIndices"
-                          :loading="data.loading">
-                </v-select>
+                <custom-v-autocomplete multiple
+                                       label="Indices"
+                                       name="Indices"
+                                       id="indices"
+                                       v-model="searchIndices"
+                                       :items="data.body | sortIndices"
+                                       :loading="data.loading">
+                  <template slot="item" slot-scope="data">
+                    <v-list-tile-action>
+                      <v-checkbox color="primary" :value="searchIndices.includes(data.item)"></v-checkbox>
+                    </v-list-tile-action>
+                    <v-list-tile-content>
+                      {{data.item}}
+                    </v-list-tile-content>
+                  </template>
+                </custom-v-autocomplete>
               </template>
             </data-loader>
           </v-flex>
@@ -53,35 +59,36 @@
 </template>
 
 <script>
-  import ResultsTable from '@/components/Browse/ResultsTable'
+  import ResultsTable from '@/components/Search/ResultsTable'
   import ReloadButton from '@/components/shared/ReloadButton'
+  import CustomVAutocomplete from '@/components/shared/CustomVAutocomplete'
 
   export default {
-    name: 'Browse',
+    name: 'Search',
     props: {
       executeSearch: {
         default: false
       }
     },
     computed: {
-      browseQ: {
+      searchQ: {
         get () {
-          return this.$store.state.browse.q
+          return this.$store.state.search.q
         },
         set (q) {
-          this.$store.commit('setBrowseQ', q)
+          this.$store.commit('setSearchQ', q)
         }
       },
-      browseIndices: {
+      searchIndices: {
         get () {
-          return this.$store.state.browse.indices
+          return this.$store.state.search.indices
         },
         set (indices) {
-          this.$store.commit('setBrowseIndices', indices)
+          this.$store.commit('setSearchIndices', indices)
         }
       },
       searchParams () {
-        return {q: this.browseQ, index: this.browseIndices}
+        return {q: this.searchQ, index: this.searchIndices}
       }
     },
     methods: {
@@ -96,7 +103,8 @@
     },
     components: {
       ResultsTable,
-      ReloadButton
+      ReloadButton,
+      CustomVAutocomplete
     }
   }
 </script>