+++ src/mixins/GetElasticsearchAdapter.js
@@ -1,28 +1,19 @@
-import ElasticsearchAdapter from '../services/ElasticsearchAdapter'
-import ConnectService from '../services/elasticsearch/ConnectService'
+import ConnectionService from '../services/elasticsearch/ConnectionService'
 
 const GetElasticsearchAdapter = {
   methods: {
     getElasticsearchAdapter () {
       if (this.$store.state.connection.elasticsearchAdapter !== null) {
-        return new Promise((resolve) => resolve(this.$store.state.connection.elasticsearchAdapter))
+        return Promise.resolve(this.$store.state.connection.elasticsearchAdapter)
       } else {
-        if (this.$store.state.connection.elasticsearchClient !== null) {
-          let adapter = new ElasticsearchAdapter(this.$store.state.connection.elasticsearchClient)
-          this.$store.commit('setElasticsearchAdapter', adapter)
-          return new Promise((resolve) => resolve(adapter))
-        } else {
-          let connectService = new ConnectService(this.$store.state.connection.elasticsearchHost)
-          return connectService.connect().then(
-            client => {
-              this.$store.commit('setElasticsearchClient', client)
-              let adapter = new ElasticsearchAdapter(client)
-              this.$store.commit('setElasticsearchAdapter', adapter)
-              return adapter
-            },
-            error => this.$store.commit('setErrorState', error)
-          )
-        }
+        let connectionService = new ConnectionService(this.$store.state.connection.elasticsearchHost)
+        return connectionService.getAdapter().then(
+          adapter => {
+            this.$store.commit('setElasticsearchAdapter', adapter)
+            return adapter
+          },
+          error => this.$store.commit('setErrorState', error)
+        )
       }
     }
   }