+++ src/mixins/IsConnected.js
@@ -1,10 +1,9 @@
 import { CONNECTION_STATES } from '../consts'
 
 const IsConnected = {
-  methods: {
+  computed: {
     isConnected () {
-      return (this.$store.state.connection.elasticsearchClient != null) &&
-        (this.$store.state.connection.status === CONNECTION_STATES.SUCCESS)
+      return this.$store.state.connection.status === CONNECTION_STATES.SUCCESS
     },
     wasConnected () {
       return this.$store.state.connection.wasConnected